simplify dcql structure
