+++ src/oid4vci/issuer/create_offer.rs
@@ -13,8 +13,8 @@ use http::StatusCode;
 
 use crate::generate;
 use crate::oauth::GrantType;
-use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::error::{invalid, server};
+use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::provider::{Metadata, Provider, StateStore, Subject};
 use crate::oid4vci::state::{Expire, Offered};
 use crate::oid4vci::types::{
