+++ src/oid4vci/issuer/credential_offer.rs
@@ -16,8 +16,8 @@
 
 use anyhow::Context as _;
 
-use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::error::invalid;
+use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::provider::{Provider, StateStore};
 use crate::oid4vci::types::{CredentialOffer, CredentialOfferRequest, CredentialOfferResponse};
