+++ src/oid4vci/issuer/statuslist.rs
@@ -2,8 +2,8 @@
 
 use anyhow::Context as _;
 
-use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::error::invalid;
+use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::provider::Provider;
 use crate::token_status::{StatusListRequest, StatusListResponse, StatusStore};
