+++ src/oid4vci/server/authorize.rs
@@ -72,8 +72,8 @@ use chrono::Utc;
 
 use crate::generate;
 use crate::oauth::GrantType;
-use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::error::{invalid, server};
+use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::provider::{Metadata, Provider, StateStore, Subject};
 use crate::oid4vci::state::{Authorized, Expire, Offered};
 use crate::oid4vci::types::{
