+++ src/oid4vci/server/register.rs
@@ -2,8 +2,8 @@
 
 use anyhow::Context as _;
 
-use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::error::server;
+use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vci::provider::{Provider, StateStore};
 use crate::oid4vci::state::Token;
 use crate::oid4vci::types::{RegistrationHeaders, RegistrationRequest, RegistrationResponse};
