+++ src/oid4vp/handlers/create_request.rs
@@ -8,8 +8,8 @@ use anyhow::Context;
 use chrono::Utc;
 
 use crate::generate;
-use crate::oid4vp::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vp::error::invalid;
+use crate::oid4vp::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vp::provider::{Metadata, Provider, StateStore};
 use crate::oid4vp::state::Expire;
 use crate::oid4vp::verifier::{
