+++ src/oid4vp/handlers/request_uri.rs
@@ -18,8 +18,8 @@ use anyhow::Context;
 use credibil_jose::JwsBuilder;
 
 use crate::oid4vp::JwtType;
-use crate::oid4vp::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vp::error::invalid;
+use crate::oid4vp::handlers::{Body, Error, Handler, Request, Response, Result};
 use crate::oid4vp::provider::{Provider, StateStore};
 use crate::oid4vp::verifier::{ClientId, RequestObject, RequestUriRequest, RequestUriResponse};
